Removed vnaconv_s2zimn from the public API
This function calculated the input impedances down the major diagonal of a
rectangular S-parameter matrix.  It really existed only as a special-case
for vnadata_convert.  It wasn't documented well, and wasn't useful enough
as a standalone function to justify the trouble of giving it proper
documentation.  Removed it from the public API by changing it to a static
function within vnadata_convert.c.