New vnacal_apply_t replaces vnacal_input_t
The vnacal_input_t structure and associated functions give incorrect
results when trying to compute the S-parameters of a DUT that has more
ports than the VNA.  It's supposed to be possible to do this by making
multiple measurements of the DUT, where in each measurement trial the
VNA is connected to different subsets of DUT ports according to a port
map, while the unconnected DUT ports are terminated with perfect loads.
The problem is that the mathematical model assumes that the DUT ports
that are not connected to the VNA in a given measurment trial see port
match reflections from VNA ports they're not connected to.  It takes
these port match terms from other measurement trials in which the DUT
port was connected to the VNA.

The current code also has the awkward restriction that a given pair of
DUT ports always has to be measured by the same pair of VNA ports in
a given measuring sequence.  This makes the port map complicated in
that it's a map of pairs to pairs rather than simply ports to ports.
Measured values must be discarded whenever they're not consistent
with the map.

This commit replaces vnacal_input_t with a new abstract data type,
vnacal_apply_t, that addresses the problems above.  Each measurement
trial adds equations to a linear system for the current port connection
arrangement.  DUT ports can be mapped arbitrarily to VNA ports on each
trial.  If more measurements are given than necessary, the new code uses
QR decomposition to find the best solution in a least squared sense.

These changes break source-level compatibility.