Fixed A matrix dimensions in vnacal_input_apply
In vnacal_input_apply, the A matrix was incorrectly given dimensions
(ports x ports) rather than (columns x columns).  This would have given
nonsense results if not for something else that was masking the bug:
the macro for accessing the A matrix used columns rather than ports for
the column dimension.  This caused the loops trying to initialize A with
the identify matrix to fill it with a different pattern.  For example
for rows = 3, columns 2, A would have been initialized with:

	1 0 1
	0 1 0
	? ? ?	<-- unitialized

(which actually is the identity matrix for 2x2 appended with some
other stuff).  But the macro definition also caused the I - E*X terms
to be placed as if the matrix had had dimension (columns x columns).
So it was working correctly, didn't overrun memory and didn't use the
unitialized matrix elements, but it was working for the wrong reasons.

Changed the dimensions of A and limits of its initializing loops to
(columns x columns).

Also merged the diagonal and off-diagonal cases of filling in the I +
E*X values below -- these become identical if we add the E*X term into
the existing identify matrix instead of replace with I + E*X.