Simplified and cleaned up test_vnacal_apply_helper
Made several simplifications and clean-ups to test_vnacal_apply_helper.
First, the dimensions of the B, S and U matrices and the x and v vectors
were all based on max_dim (number of ports) rather than their natural
dimensions.  This required a lot of unnecessary copying and expanding
of matrices.  Eliminated most of these and a lot of unnecessary code,
keeping only a (rows x rows) matrix A, and rows-length column vectors
b and x.  Added several comments to make the code easier to understand.

Changed test_vnacal_input_apply to not skip the remaining tests if it
skips an earlier test due to a singular matrix.  Added a test for the
2x2 matrix case.

Elsewhere in the file, renamed a few functions and variables.  Added error
term column headers under -v.  Added comments.