Modified _vnacal_rfi to allow extrapolation
The library allows a small amount of extrapolation if the frequency
range we're extrapolating from falls just slightly out of bounds at
the endpoints.  Before, the library used zero-order extrapolation for
this, i.e. it simply extended the first and last points.  Modifed the
code to let the function interpolation code estimate the actual value
beyond the endpoint.