Removed inner const in vnacal_new*, vnacal_apply* APIs
In the functions that take pointer to matrix of pointer to value inputs,
e.g. vnacal_new_add_* and vnacal_apply*, both pointers in the formal
argument list were declared const indicating that the function would
neither change the matrix of pointers nor the vectors it pointed to.
But C makes multiple levels of const inconvenient because you cannot
pass a non-const matrix of non-const elements to a function expecting
a pointer to const matrix of const elements -- only the first level
of const is allowed.  As a result, several places in the code, we were
having to add intermediate variables or type casts to pass the correct
argument type, pointing out that it was more of a misfeature than not.
Removed the inner const declarations to better match the common use case.