Removed negative indents in vnaconv.3
Instead of negative indents, changed to just break the line sooner in
nroff mode.  Fixed a couple old references to Link with -lvnaconv --
it's just -lvna.

Eliminated the internal vnaconv_array2_t typedef, replacing it with
the full declaration written out so it's more clear that code and
documentation are the same.

Except for man page formatting, no functional changes