Reorganized internal test library
Renamed the internal test library from libvnacaltest.a to libt.a and
added in the common libt.c (was test.c) module to make it a generic
test library for all tests rather than specific to the vnacal tests.
Gave all files and global symbols a "libt" prefix to make it more clear
where they come from.

No functional changes.