Added test-vnacal-Van-hamme
Add test for the iterative autocalibration solution described in Flexible
Vector Network Analyzer Calibration With Accuracy Bounds Using an 8-Term
or a 16-Term Error Correction Model.  Currently, this algorithm is finding
less-precise solutions than the initial guesses and the test compare
tolerance is set pretty high.  But we're not yet weighting equations
correctly according to the paper, and there's hope for improvement.
Until then, this test still serves as a regression test protecting
against changes that make the result even worse.