Added vnaproperty to the documented API
Re-worked and documented the vnaproperty functions, enhancing the
interface to support list insert and append, creation of empty maps
and lists, the ability to work on subtrees and YAML import/export.

Added vnacal_property_get_subtree and vnacal_property_set_subtree
functions so that vnacal properties can be managed using the general
interface.