Moved VPT_T enum up; reserved slot for VPT_U
Re-ordered the members of the vnadata_parameter_type_t enum to group T
parameters with S parameters (the scattering parameters) as opposed to
grouping them with the other 2x2 parameters.  Reserved a slot after T
for future addition of inverse T ("U" parameters).

This is fundamentally a cosmetic change: it doesn't introduce any
functional differences.  But it does break binary compatibility --
application code must be recompiled.