Added U parameters (inverse T)
Added inverse scattering-transfer (U parameters) to vnaconv and vnadata.
The U parameters for a cascade of two-port networks is the right-to-left
matrix product of the U parameters of each successive stage.  The pair
T-U follows the same pattern of the other complements: Z-Y, H-G and A-B.

We use U parameters already in vnacal to solve for error terms when
there are more RF detectors than RF sources, e.g. in classic half
s-parameter SOLT.  In the vnacal usage, it's a little different in that
the elements of the U matrix are themselves matrices; nevertheless,
the example illustrates the utility of U parameters.

Another application: when analyzing a cascade of stages, if the
S parameters of a particular stage have s21 == 0 and s12 != 0, the
stage cannot be represented in T parameters but can be represented in
U parameters.