Disallow rectangular S parameter matrices
Until now, the library has permitted creation of rectangular S-parameter
matrices.  The intent was to try to accommodate simple VNAs that measure
only s11 and s21.  But in practice, rectangular S-parameter matrices
aren't very useful.  First, s11 and s21 can be measured by themselves
only if we know that the port match error term at VNA port 2 is zero,
i.e. if we drop one of the error terms; or alternatively, if we know
that the the DUT is fully symmetrical: s11 == s22 and s12 == s21.
Vnacal doesn't produce rectangular S matrices.  In the earlier case,
it doesn't support dropping the port match term, and in the later case,
it simply produces the full matrix.

Touchstone file formats also don't support rectangular matrices.
Version 2, supports upper and lower triangular matrices, though the
library converts these to their square counterparts when loaded.
The higher memory usage that results from storing the redundant terms
is offset by a simpler and leaner code.

Finally, if we do manage to create a rectangular S matrix, there aren't
many operations we can do on it.  It cannot be converted to any of the
other matrix parameter types: Z, Y, H, etc.  All we can do is find the
return loss, VSWR, Zin, and RC and RL equivalents on the major diagonal
terms. Or we can load/save it in NPD format.

This change requires S matrices to be square.