Reset vnacal file version to 1.0 with compatibility
During development, the VNA calibration file format changed several times.
The original 1.x versions used a custom format.  We switched to YAML in
2.x, dropping support for 1.x.  Like the 1.x versions, however, the 2.x
versions supported only E12 error terms.  In 3.x, we generalized the error
term format, adding support for several variations of T and U error terms.

This commit resets the version to 1.0 while still providing compatibility
with the older 2.x and 3.x versions.  In the older versions, the file
began with a #VNACAL <major>.<minor> header line.  The new header
line lower-cases the AL to #VNACal <major>.<minor> so that there's no
ambiguity with the older format.  Now we internally map the old 2.x to
0.2 and the old 3.x to 1.0.  Both can still be read.