Fix vnadata_save return loss for one-port DUT
The vnadata_save function gave an error if you tried to save return loss
in the single-port (reflection only) case.  The error message reported
that off-diagonal S parameters were required for this format.  But that
check was intended for insertion loss, not return loss.  Moved the check
to the right place.