Allow a certain percentage of test-vnacal-Van-hamme failures
Because we're adding random measurement errors, we can expect the test
to fail with a certain probability.  Allow up to 2% of the test trials
to fail to protect against spurious test failure.

Now that we have a bit of margin, we also add error to the accuracy of
the short and open standards in addition to the measurement repeatibilty
errors we were already adding.