Change directed reachability matrix to undirected connectivity matrix
In vnacal_new_add_common, we analyze the S parameter matrix of each
calibration standard to determine which pairs of VNA ports have a signal
path through the standard, as opposed to leakage in the VNA or test set.
This is used to identify leakage terms outside of the linear system and
to eliminate zero equals zero equations from the linear system.

We were using Floyd Warshal closure to generate a directed graph of
the reachability.  For example, if the standard passed signal from
port 1 to port 2, but perfectly blocked it in the opposite direction,
we considered port 2 reachable from 1, but 1 not reachable from 2.
Thus, we could opportunistically use S12 as a leakage measurement.

But as we prepare to introduce the Van hamme "V" matrix that maps the
residuals of the equations for a standard back to measurement errors,
we need a new definition of connectivity between ports.  The reason is
that the V matrix is the inverse of a matrix that may contain unknown
elements, and an element of V is valid only if we can show that there
are no unknown elements in the same diagonal block of the matrix, if
the matrix were to be rearranged into block diagonal form.  For this,
we need a definition of connectivity that indicates if two VNA port
pairs would fall in the same diagonal block of V.

We considered keeping the directed form of reachability for determining
leakage terms, and introducing the undirected form alongside it for
determining if an element of the V matrix is valid.  But there's not
much practical difference between these.  They're different only if
you're using a perfect bridge or circulator as a calibration standard,
and then the only difference is whether or not we can opportunistically
add a leakage measurement for the blocked direction.  For simplicity,
we've decided to simply drop the older definition in favor of the new.