Add V matrix support to vnacal_new_add_common
Split vnacal_new_build_equation.c out of vnacal_new_add_common.c and
enhance it to include multiplication by the V matrix.  Multiplication
by the V matrix transforms the residuals of the equations for a
measured standard into errors in each of the measurements.  This is
valuable when we're given the error model of the measurements (via
vnacal_new_set_m_error) in that it lets us normalize the errors and test
if they're consistent with the error model.

Add a new constraint that if measurement errors are set with
vnacal_set_m_error and the parameter type is T16 or U16, that a full
S parameter matrix must be given with each calibration measurement,
i.e. all VNA ports must be connected to standards.

Because vnacal_new_set_m_error is not yet a documented function, there
are no functional changes to the currently documented API.