Rename vnacal_new_coefficient_t to vnacal_new_term_t
The vnacal_new_coefficient_t structure is named as it is because it
contributes one element to the coefficient matrix used to solve the
error terms.  But at the same time, it represents an expanded term of
one of the equations used to create that matrix.

In the next commit, we plan to introduce a weighting matrix, V,
that transforms the residuals of the equations to measurement errors.
Integrating the V matrix into the equations, however, breaks the 1:1
relationship between equation terms and elements in the coefficient
matrix: more than one term can contribute to the same matrix element.
This rename is in preparation for that change.

The commit changes the struct member prefix from "vnc" to "vnt", changes
"vs_next_coefficient" of the vnacal_new_solve equation iterator to
"vs_next_term", and also contains a few other related changes.

No functional changes: internal renames only.