Fix sign error in vnacal_new_solve_auto
Change the sign of k_vector and d_vector so that we can subtract
d_vector from p vector instead of adding to make the correction to be
more consistent with Gauss-Newton.  This commit also fixes a bug in the
existing code: the k_vector values were already inverted.

This commit improves accuracy when correlated parameters are present.