Re-work vnacal_new_solve_auto to use V matrices
Change the equation weighting approach in vnacal_new_solve_auto to use V
matrices as described in the Van hamme paper instead of the per-equation
weighting scheme we were using before.  This corrects an important
deviation from the target algorithm.  With the change, we see about an
8-9% improvement in median and average RMS error, and some flattening
of the right tail in test-vnacal-Van-hamme.