Add test-vnacal-v-matrices
New test: test that the V matrices correctly convert the residuals of the
equations of a measured standard to errors in each associated measurement.

Strategy: add a series of random standards to vnacal_new with errors
added into the measurements.  Use more standards than needed so that that
the linear system is overdetermined and the equations produce residuals.
Set up the environment for vnacal_new_solve, but don't actually solve.
Instead, use the known error terms generated by the test to update
the V matrices.  Compute the residuals of each equation and check that
they're equal to the measurement errors we added above.