Replace calc_rms_error with vs_calc_pvalue
When a measurement error model is given via vnacal_new_set_m_error, we've
been calculating the RMS error of the residuals as a way to validate that
the measurements are consistent with the error model.  A couple problems
with this are: (1) that the code assumes the residuals are already in
units of measurement error, which they're not -- we need to multiply
by the V matrix to convert them to units of measurement error, and (2)
the resulting metric isn't a very meaningful value -- it requires the
user to come up with an arbitrary constant to decide if the residuals
are too large.

This commit replaces calc_rms_error with a new method, vs_calc_pvalue,
that's more rigorously rooted in statistics.  It uses the V matrices to
convert the residuals of the equations to units of measurement error,
calculates their chi-squared statistic, then finds the p-value that the
residuals we observe are explained by the error model.