Support measurement error modeling in vnacal_new_solve_simple
Enhance vnacal_new_solve_simple to weight the equations using the V matrix
when the system is overdetermined and a measurement error model is given
via vnacal_new_set_m_error.  In this mode, the solver iterates until
the V matrix converges (or fails to converge) weighting the equations
consistently with the error model.