Add analytic TRL calibration method
This commit adds an analytic calibration method for 2x2 T8, U8, TE10
and UE10 with three standards: through, unknown symmetrical double
reflect, and unknown symmetrical matched line; and no measurement error
modeling.  When it applies, this method is much faster than the general
auto calibration method.  It achieves close to full precision accuracy
without having to iterate.

The library automatically detects this case and applies it when possible.
There are no changes to the library API.