version 0.3.4
The main externally visible feature of this release is a new
analytic TRL solver for 2x2 T8, U8, TE10 and UE10 error term types
that's significantly faster than the general autocalibration method.
It achieves close to full double floating precision without iteration.
The new solver is applied automatically when it can be -- no changes
are needed in programs using the library to take advantage of it.

Most of the other changes are internal code reorganization, mostly
related to VNA calibration measurement error modeling.  This feature
is almost ready to be documented into the public API.