Fix copying unitialized p_vector when no iteration
In vnacal_new_solve_auto, in the special case that the system converges on
the first iteration, i.e. unknown parameters are equal to their initial
guesses and error terms are "perfect", we were breaking the loop and
copying p_vector from the "best" solution, which at this point had never
been saved.  This commit adds an up_to_date flag to suppress the copy
if the current parameters are already the ones we want.