Use array parameter notation the the 2x2 functions of vnaconv.3
For the 2x2 conversion functions, changed the function parameter
declaration style in the synopsis from the more correct (from a pointer
arithmetic perspective) pointer to element type form, e.g. double complex
(*s)[2] to the more intuitive (for people coming from other languages)
array form, e.g. double complex s[2][2].  The intent is to make it more
clear what the user is expected to pass to the function.  Added a notes
section with an explanation.  The actual code continues to use the honest
type style.