Improve analytic TRL solution
If the measurements fit the TRL model, then ml12 * mt21 == ml21 * mt12.
In the solution for L, a was the first and c was the second.  Now, both
a and c are the average of the two.

Added a singularity check to the L solution (if a and c are zero).

Replaced the R solution with a simpler, more symmetric form.