Allow a small number of random solve failures in test-vnacal-pvalue
When testing with random measurement errors, a small percentage of test
cases will fail in the solve step.  This changes the test to tolerate up
to 3 failures out of every 1000 to avoid failing the whole test (which
is really meant to test the p-value distribution rather than success
or failure of individual cases).  Fixed an off-by-one array overrun in
the test.  Made the KS statistic cut-off more correct and commented how
to select the constant.

Fixed a couple unused variables and updated comments in other tests.