Avoid a few compiler warnings
Under more strict warning options than -Wall, gcc complained that zero
length variable arrays might not be initialized.  Zero length VLA's
aren't really permitted under C99, so changed to code to avoid the ones
it complained about.  Also fixed a warning that strncpy was copying into a
buffer the same size as the limit.  This practice isn't really illegal --
when using the strn* functions, if the limit is reached, the string is not
expected to be NUL-terminated.  Still, it was easy to make the buffer one
byte larger.  Avoiding the non-NUL terminated string makes debug simpler
in that you can just print the string in all cases.  Finally, avoided
the POSIX constant M_PI in example programs that already were defining PI.