Provide better random number generator fallback
If the "random" function (typically an implementation of the Mersene
twister algorithm) is not available, then archdep.h would fall back
to using "rand" (typically a 15 bit LCG algorithm).  The problem with
the 15 bit random number generator, though, is it too often creates
"inconvenient" values that cause the tests to fail.  This commit adds
a better alternative to the 15 bit generator.

Also, we were using the constant RAND_MAX with random, which is
not correct because RAND_MAX is specific to rand.  In Linux, rand is
implemented using random, so it happens to be the correct constant but
for the wrong reason.  This commit fixes that by introducing RANDOM_MAX
(if not already defined) and using that instead.