Make the use of RANDOM_MAX more precise.
When using the Box Muller method to create two normally distributed random
real numbers, we need two uniformly distributed random numbers, one in
the range (0..1] and the other in the range [0..1).  These bounds were
not correct.  Also, when choosing a random angle, we want the [0..1) form
to avoid repeating the zero angle point.