Fix incorrect comments describing UE14 S to M calculation
This commit corrects a mistake in the comments in vnacal_layout.h
describing how to calculate the measurement matrix from S-parameter matrix
of the standards and the UE14 error parameters.  The description of how
to build the `B` matrices was wrong.  Changed it to match the way it's
coded in the library.  Verified using Mathematica that when the new result
is plugged into the M->S equation, it correctly recovers the S parameters.