Fix vnacal_new_solve_auto convergence problem
When using measurement error modeling with unknown parameters, the
system could fail to converge due to a bug in _vnacal_new_solve_auto().

We were updating the "v" matrices (that scale the equations such that
the error in the residuals due to measurement error is 1) in between
calculating the error parameters (x_vector) and calculating the change in
the unknown parameters (p_vector).  But this caused inconsistent scaling,
even scaling by negative values in the change in the unknown parameters,
causing them to diverge in some cases.  This change moves the update
of the v matrices to the top of the loop where it was in the Van hamme
paper and significantly improves stability.