Fix _vnacal_new_solve_auto stop condition
Before, _vnacal_new_solve_auto acted on best or non-best solution in the
middle of the loop.  It saved or restored x vector, p vector, j matrix,
k vector and the v matrices (but not the q2^H matrix).  And it checked if
the changes in x vector and p vector were small enough to stop iteration
-only- after finding a best solution.  Saving so much state made the
code unnecessarily complicated, and checking for convergence only after
a best solution caused convergence problems if we landed on the minimum
and could then never find a better solution.

This change simplifies the code so that we calculate the new x vector
and d vector (proposed change in p vector), and save only x vector and p
vector if we have a best solution, or restore them if not.  Now, we always
check for convergence, even on a non-best solution.  And we no longer
penalize the convergence check by multiplying by marquardt_multipler:
if the marquardt parameter causes the change in the parameters to become
small enough to meet our stop condition and we haven't found a better
solution, then we must already be at a minimum.