Warn if loaded .s2p file doesn't have two ports
When vnadata_load loads Touchstone 1 files, it generates a warning
(but otherwise works) if the number of ports in the file doesn't match
the number in the filename.  Except that we treated .s2p as wild and
suppressed the warning.  This commit removes the exception so that the
warning is generated also if a .s2p file has other than two ports.