Add vnadata_set_name, vnadata_get_name
The  vnadata_get_name()  and  vnadata_set_name()  functions  provide  a
mechanism  to  get and set a meaningful name for a vnadata_t structure,
such as the name of the device described by the network parameter data.
This  name  is  included in some error messages.  If vnadata_set_name()
has not been called, vnadata_get_name() returns a default name  derived
from  the  filename used in the most recent load or save operation, or,
if no filename is available, from the dimensions of the data matrix.