Rename gamma parameter to coefficient when it's not reflect only
The vnacal_make_scalar_parameter and vnacal_make_vector_parameter
functions called their coefficients gamma and gamma_vector, respectively.
But these can be either coefficients of reflection or transmission,
and in the later case, gamma is not the right name.  Changed the
parameter names to coefficient and coefficient_vector to be more clear.
Changed internal variables to match.  Because this changes the name of
a formal parameter only, there's no compatibility break.