Simplify vnadata_save file type logic
The vnadata_save and vnadata_fsave functions had logic that intended to
write files with an extension of .ts in Touchsone 1 format if there were
four or fewer ports and per-port reference impedances were not in-use.
Otherwise, it would write in Touchstone 2.  This would have been
confusing behavior and would have provided no way for the user to force
V2 format to .ts files.  Fortunately, however, there was a bug in the
code that caused it to always write to .ts files in Touchstone 2 format.
This commit removes the broken logic, keeping the semantics as they are.