Change vnacal_get_z0_vector and vnacal_delete_parameter_matrix
This commit fixes a couple API consistency problems introduced in 0.5.0.
First, the signature of vnacal_get_z0_vector is changed to move the
output vector to the end of the argument list for consistency with
other functions.  Second, vnacal_delete_parameter_matrix now takes a
writable rather than const pointer to the parameter matrix and returns
int instead of void.  It marks the entries as -1 as it deletes them.
Both of these are compatibility breaks, but on just-added functions only.